Social share initial implementation (hidden for now)

+++ FlipCard-WP/RandomGame.xaml.cs
@@ -12,6 +12,7 @@ using System.Windows.Input;
 using System.Diagnostics;
 using System.Threading;
 using System.Windows.Media.Animation;
+using Microsoft.Phone.Tasks;
 
 namespace FlipCard_WP
 {
@@ -36,6 +37,10 @@ namespace FlipCard_WP
         int step = 0;
         Random rgn = new Random();
         bool PBegin = false;
+
+        int playerOverall = 0;
+        int cpuOverall = 0;
+        bool shared = false;
 
         public RandomGame()
         {
@@ -398,6 +403,22 @@ namespace FlipCard_WP
             img.Visibility = System.Windows.Visibility.Collapsed;
         }
 
+        private void sharestatus_Click(object sender, EventArgs e)
+        {
+            ShareStatusTask shareStatusTask = new ShareStatusTask();
+            String cpuwin = " CPU is beating me :(";
+            String playerwin = " I'm winning! :D";
+            String tie = " We're even! :)";
+            String res = playerOverall.ToString() + " " + cpuOverall.ToString();
+            if (cpuOverall == playerOverall)
+                shareStatusTask.Status = "First twitter message from #FlipCard! " + res + tie;
+            else
+                shareStatusTask.Status = "First twitter message from #FlipCard! " + res + (cpuOverall > playerOverall ? cpuwin : playerwin);
+            this.shared = true;
+            shareStatusTask.Show();
+        }
+
+
         private void PhoneApplicationPage_BackKeyPress(object sender, System.ComponentModel.CancelEventArgs e)
         {
             MessageBoxResult Result = MessageBox.Show("Go back?", "Exit menu", MessageBoxButton.OKCancel);
@@ -410,20 +431,22 @@ namespace FlipCard_WP
         protected override void OnNavigatedTo(NavigationEventArgs e)
         {
             String id = "";
-            if(NavigationContext.QueryString.TryGetValue("id",out id)){
-               if(NavigationService.CanGoBack) {
+            String cpuScore = "";
+            String playerScore = "";
+
+            if (NavigationContext.QueryString.TryGetValue("player", out playerScore))
+                this.playerOverall = int.Parse(playerScore);
+
+            if (NavigationContext.QueryString.TryGetValue("cpu", out cpuScore))
+                this.cpuOverall = int.Parse(cpuScore);
+
+            if (NavigationContext.QueryString.TryGetValue("id", out id))
+            {
+                if (NavigationService.CanGoBack && !this.shared)
+                {
                     NavigationService.RemoveBackEntry();
                 }
             }
-
-
-            //MessageBox.Show("Sono qua");
-            //if(e.Uri==(new Uri("/RandomGame.xaml", UriKind.Relative))){
-            //    MessageBox.Show("Sono proprio qua");
-            //    NavigationService.RemoveBackEntry();
-             // }   
-                
-            
         }
 
         private void CardOnTable_Completed(object sender, EventArgs e)
